fix: modal error
