+++ src/components/ClubMatchResults.js
@@ -1,222 +1,205 @@
 import React, { useState } from "react";
-import { Carousel, Container, Modal } from "react-bootstrap"
+import { Carousel, Container } from "react-bootstrap"
 import Opponent from "./Opponent"
-import OpponentDetail from "./OpponentDetail";
 
 const club_match_results = [
-    {
-        club_id: 1,
-        opponent_name: "鹿島アントラーズ",
-        win: 5,
+  {
+    club_id: 1,
+    opponent_name: "鹿島アントラーズ",
+    win: 5,
+    lose: 0,
+    draw: 1,
+    total_match_num: 100,
+    winning_rate: 99.0,
+    invincibility_rate: 100.0,
+    detail: {
+      home_result: {
+        win: 33,
         lose: 0,
         draw: 1,
-        total_match_num: 100,
-        winning_rate: 99.0,
+        winning_rate: 97.05,
         invincibility_rate: 100.0,
-        detail: {
-            home_result: {
-                win: 33,
-                lose: 0,
-                draw: 1,
-                winning_rate: 97.05,
-                invincibility_rate: 100.0,
-            },
-            away_result: {
-                win: 33,
-                lose: 0,
-                draw: 0,
-                winning_rate: 100.0,
-                invincibility_rate: 100.0,
-            },
-            neutral_result: {
-                win: 33,
-                lose: 0,
-                draw: 0,
-                winning_rate: 100.0,
-                invincibility_rate: 100.0,
-            }
-        }
-    },
-    {
-        club_id: 2,
-        opponent_name: "ガンバ大阪",
+      },
+      away_result: {
+        win: 33,
+        lose: 0,
+        draw: 0,
+        winning_rate: 100.0,
+        invincibility_rate: 100.0,
+      },
+      neutral_result: {
+        win: 33,
+        lose: 0,
+        draw: 0,
+        winning_rate: 100.0,
+        invincibility_rate: 100.0,
+      }
+    }
+  },
+  {
+    club_id: 2,
+    opponent_name: "ガンバ大阪",
+    win: 49,
+    lose: 49,
+    draw: 10,
+    total_match_num: 108,
+    winning_rate: 45.37,
+    invincibility_rate: 54.62,
+    detail: {
+      home_result: {
         win: 49,
+        lose: 0,
+        draw: 1,
+        winning_rate: 98.0,
+        invincibility_rate: 100.0,
+      },
+      away_result: {
+        win: 0,
         lose: 49,
-        draw: 10,
-        total_match_num: 108,
-        winning_rate: 45.37,
-        invincibility_rate: 54.62,
-        detail: {
-            home_result: {
-                win: 49,
-                lose: 0,
-                draw: 1,
-                winning_rate: 98.0,
-                invincibility_rate: 100.0,
-            },
-            away_result: {
-                win: 0,
-                lose: 49,
-                draw: 1,
-                winning_rate: 0.0,
-                invincibility_rate: 2.0,
-            },
-            neutral_result: {
-                win: 0,
-                lose: 0,
-                draw: 8,
-                winning_rate: 0.0,
-                invincibility_rate: 100.0,
-            }
-        }
-    },
-    {
-        club_id: 3,
-        opponent_name: "北海道コンサドーレ札幌",
-        win: 2,
-        lose: 1,
+        draw: 1,
+        winning_rate: 0.0,
+        invincibility_rate: 2.0,
+      },
+      neutral_result: {
+        win: 0,
+        lose: 0,
+        draw: 8,
+        winning_rate: 0.0,
+        invincibility_rate: 100.0,
+      }
+    }
+  },
+  {
+    club_id: 3,
+    opponent_name: "北海道コンサドーレ札幌",
+    win: 2,
+    lose: 1,
+    draw: 2,
+    total_match_num: 5,
+    winning_rate: 40.0,
+    invincibility_rate: 60.0,
+    detail: {
+      home_result: {
+        win: 1,
+        lose: 0,
         draw: 2,
-        total_match_num: 5,
-        winning_rate: 40.0,
-        invincibility_rate: 60.0,
-        detail: {
-            home_result: {
-                win: 1,
-                lose: 0,
-                draw: 2,
-                winning_rate: 33.33,
-                invincibility_rate: 33.33,
-            },
-            away_result: {
-                win: 1,
-                lose: 1,
-                draw: 0,
-                winning_rate: 50.0,
-                invincibility_rate: 50.0,
-            },
-        }
-    },
-    {
-        club_id: 4,
-        opponent_name: "柏レイソル",
+        winning_rate: 33.33,
+        invincibility_rate: 33.33,
+      },
+      away_result: {
+        win: 1,
+        lose: 1,
+        draw: 0,
+        winning_rate: 50.0,
+        invincibility_rate: 50.0,
+      },
+    }
+  },
+  {
+    club_id: 4,
+    opponent_name: "柏レイソル",
+    win: 3,
+    lose: 0,
+    draw: 0,
+    total_match_num: 3,
+    winning_rate: 100.0,
+    invincibility_rate: 100.0,
+    detail: {
+      home_result: {
         win: 3,
         lose: 0,
         draw: 0,
-        total_match_num: 3,
         winning_rate: 100.0,
         invincibility_rate: 100.0,
-        detail: {
-            home_result: {
-                win: 3,
-                lose: 0,
-                draw: 0,
-                winning_rate: 100.0,
-                invincibility_rate: 100.0,
-            },
-        }
-    },
-    {
-        club_id: 5,
-        opponent_name: "サガン鳥栖",
+      },
+    }
+  },
+  {
+    club_id: 5,
+    opponent_name: "サガン鳥栖",
+    win: 2,
+    lose: 8,
+    draw: 0,
+    total_match_num: 10,
+    winning_rate: 20.0,
+    invincibility_rate: 20.0,
+    detail: {
+      home_result: {
         win: 2,
-        lose: 8,
+        lose: 2,
         draw: 0,
-        total_match_num: 10,
-        winning_rate: 20.0,
-        invincibility_rate: 20.0,
-        detail: {
-            home_result: {
-                win: 2,
-                lose: 2,
-                draw: 0,
-                winning_rate: 50.0,
-                invincibility_rate: 50.0,
-            },
-            away_result: {
-                win: 0,
-                lose: 5,
-                draw: 0,
-                winning_rate: 0.0,
-                invincibility_rate: 0.0,
-            },
-            neutral_result: {
-                win: 0,
-                lose: 3,
-                draw: 0,
-                winning_rate: 0.0,
-                invincibility_rate: 0.0,
-            }
-        }
-    },
-    {
-        club_id: 6,
-        opponent_name: "上海上港",
+        winning_rate: 50.0,
+        invincibility_rate: 50.0,
+      },
+      away_result: {
+        win: 0,
+        lose: 5,
+        draw: 0,
+        winning_rate: 0.0,
+        invincibility_rate: 0.0,
+      },
+      neutral_result: {
+        win: 0,
+        lose: 3,
+        draw: 0,
+        winning_rate: 0.0,
+        invincibility_rate: 0.0,
+      }
+    }
+  },
+  {
+    club_id: 6,
+    opponent_name: "上海上港",
+    win: 1,
+    lose: 1,
+    draw: 2,
+    total_match_num: 4,
+    winning_rate: 25.0,
+    invincibility_rate: 75.0,
+    detail: {
+      home_result: {
         win: 1,
+        lose: 0,
+        draw: 1,
+        winning_rate: 50.0,
+        invincibility_rate: 100.0,
+      },
+      away_result: {
+        win: 0,
         lose: 1,
-        draw: 2,
-        total_match_num: 4,
-        winning_rate: 25.0,
-        invincibility_rate: 75.0,
-        detail: {
-            home_result: {
-                win: 1,
-                lose: 0,
-                draw: 1,
-                winning_rate: 50.0,
-                invincibility_rate: 100.0,
-            },
-            away_result: {
-                win: 0,
-                lose: 1,
-                draw: 1,
-                winning_rate: 0.0,
-                invincibility_rate: 50.0,
-            },
-        }
-    },
+        draw: 1,
+        winning_rate: 0.0,
+        invincibility_rate: 50.0,
+      },
+    }
+  },
 ]
 
 function ClubMatchResults() {
-    const [index, setIndex] = useState(0);
+  const [index, setIndex] = useState(0);
 
-    const [show, setShow] = useState(false);
-    const handleClose = () => setShow(false);
-    const handleShow = () => setShow(true);
+  const handleSelect = (selectedIndex, e) => {
+    setIndex(selectedIndex);
+  }
 
-    const handleSelect = (selectedIndex, e) => {
-        setIndex(selectedIndex);
-    }
-
-    return (
-        <Container>
-            <Carousel
-                activeIndex={index}
-                onSelect={handleSelect}
-                slide={false}
-                interval={null}
-                indicators={false}
-                prevIcon={<span aria-hidden="false" className="carousel-control-prev-icon" />}
-                nextIcon={<span aria-hidden="false" className="carousel-control-next-icon" />}
-            >
-                {club_match_results.map(club_match_result => (
-                    <Carousel.Item variant="link text-secondary button_link" onClick={handleShow}>
-                        <Opponent club_match_result={club_match_result} />
-                    </Carousel.Item>)
-                )
-                }
-            </Carousel>
-            <Modal show={show} onHide={handleClose}>
-                <Modal.Header closeButton></Modal.Header>
-                {club_match_results.map(club_match_result => (
-                    <Modal.Body>
-                        <OpponentDetail club_match_result={club_match_result} />
-                    </Modal.Body>
-                )
-                )
-                }
-            </Modal>
-        </Container>
-    );
+  return (
+    <Container>
+      <Carousel
+        activeIndex={index}
+        onSelect={handleSelect}
+        slide={false}
+        interval={null}
+        indicators={false}
+      >
+        {club_match_results.map(club_match_result => (
+          <Carousel.Item>
+            <Opponent club_match_result={club_match_result} />
+          </Carousel.Item>)
+        )
+        }
+      </Carousel>
+    </Container>
+  );
 }
 
 export default ClubMatchResults;
